added a compose file in ignore, renamed folder to trial_document_search
