+++ trial_document_search/services/similar_trail_documents_reterival_service.py
@@ -1,5 +1,5 @@
 from trial_document_search.utils.logger_setup import logger
-from database.trial_analysis.job_status import create_empty_job, add_job
+from database.trial_document_search.job_status import create_empty_job, add_job
 from trial_document_search.utils.similar_trial_documents_utils.fetch_similar_document_using_pinecone import fetch_similar_documents_using_pinecone
 from trial_document_search.utils.similar_trial_documents_utils.combine_and_ensure_unique_documents import combine_and_ensure_unique_documents
 from trial_document_search.utils.similar_trial_documents_utils.filter_documents import filter_documents
