+++ trial_document_search/utils/similar_trial_documents_utils/store_and_return_empty_response.py
@@ -1,4 +1,4 @@
-from database.trial_analysis.store_similar_trials import store_similar_trials
+from database.trial_document_search.store_similar_trials import store_similar_trials
 from trial_document_search.utils.logger_setup import logger
 
 def store_and_return_empty_response(user_data: dict, user_inputs: dict, final_response: dict) -> None:
